Reorganize embed wizard into three steps
